feat: integracao entre produto e categoria, adicionado extensao de permissao no docker compose para utilizar no ambiente de dev fedora.

+++ src/main/resources/categoria.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS categoria (
-                                         id_categoria INT AUTO_INCREMENT PRIMARY KEY,
+                                         id_categoria BIGINT AUTO_INCREMENT PRIMARY KEY,
                                          nome VARCHAR (45) NOT NULL,
     descricao VARCHAR (150)
     );
